Don't repeat analyses

+++ deploy/roles/sparkmaster/files/run_analysis.sh
@@ -11,15 +11,20 @@ MAIN=org.fhcrc.matsen.startree.spark.StarTreeSpark
 MASTER=spark://$(GET http://169.254.169.254/latest/meta-data/public-hostname):7077
 
 for f in *.bam; do
-  java -Djava.library.path=/usr/local/lib \
-    -cp $SPARK_JAR:$STAR_JAR \
-    $MAIN \
-    $MASTER \
-    --jar-path $STAR_JAR \
-    --executor-memory 4g \
-    --sample \
-    --parallelism 128 \
-    --bucket startreerenaissance \
-    --prefix $(basename $f .bam) \
-     04-A-M_IGHV_model.json ighvdj.fasta $f
+  BASE=$(basename $f .bam)
+  if [ -f ${BASE}IGHV*json.gz ]; then
+    echo "$f appears done."
+  else
+    java -Djava.library.path=/usr/local/lib \
+      -cp $SPARK_JAR:$STAR_JAR \
+      $MAIN \
+      $MASTER \
+      --jar-path $STAR_JAR \
+      --executor-memory 4g \
+      --sample \
+      --parallelism 128 \
+      --bucket startreerenaissance \
+      --prefix $(basename $f .bam) \
+       04-A-M_IGHV_model.json ighvdj.fasta $f
+  fi
 done
